Added setFlashMessage and getFlashMessage methods to the Framework\Session conitinue in 46:41 in video

+++ App/controllers/ListingController.php
@@ -4,6 +4,8 @@ namespace App\Controllers;
 
 use Framework\Database;
 use Framework\Validation;
+use Framework\Session;
+use Framework\Authorization;
 
 class ListingController
 {
@@ -19,7 +21,7 @@ class ListingController
     public function index()
     {
 
-        $listings = $this->db->Query('SELECT * FROM listings')->fetchAll();
+        $listings = $this->db->Query('SELECT * FROM listings ORDER BY created_at DESC')->fetchAll();
 
         loadView('listings/index', ['listings' => $listings]);
     }
@@ -72,7 +74,7 @@ class ListingController
 
         $newListingData = array_intersect_key($_POST, array_flip($allowedFields));
 
-        $newListingData['user_id'] = 1;
+        $newListingData['user_id'] = Session::get('user')['id'];
 
         $newListingData = array_map('sanitize', $newListingData);
 
@@ -117,6 +119,8 @@ class ListingController
 
             $this->db->query($query, $newListingData);
 
+            Session::setFlashMessage('success_message', 'Listing created successfully');
+
             redirect('/listings');
         }
     }
@@ -136,15 +140,21 @@ class ListingController
         ];
 
         $listing = $this->db->query('SELECT * FROM listings WHERE id = :id', $params)->fetch();
-
+        //check if listing exists
         if (!$listing) {
             ErrorController::notFound('Listing not found');
             return;
         }
 
-        $this->db->query('DELETE FROM listings WHERE id = :id', $params);
+        //Authorization
+        if (!Authorization::isOwner($listing->user_id)) {
+            Session::setFlashMessage('error_message', 'You are not authorized to delete this listing');
+            return redirect('/listings/' . $listing->id);
+        }
 
-        $_SESSION['success_message'] = 'Listing deleted successfully';
+        $this->db->query('DELETE FROM listings WHERE id = :id', $params);
+        //Set flash message
+        Session::setFlashMessage('success_message', 'Listing deleted successfully');
 
         redirect('/listings');
     }
@@ -162,6 +172,13 @@ class ListingController
             ErrorController::notFound('Listing not found');
             return;
         }
+
+        //Authorization
+        if (!Authorization::isOwner($listing->user_id)) {
+            $_SESSION['error_message'] = 'You are not authorized to update this listing';
+            return redirect('/listings/' . $listing->id);
+        }
+
         loadView('listings/edit', [
             'listing' => $listing
         ]);
@@ -187,6 +204,12 @@ class ListingController
         if (!$listing) {
             ErrorController::notFound('Listing not found');
             return;
+        }
+
+        //Authorization
+        if (!Authorization::isOwner($listing->user_id)) {
+            $_SESSION['error_message'] = 'You are not authorized to update this listing';
+            return redirect('/listings/' . $listing->id);
         }
 
         $allowedFields = [
